+++ android/src/main/kotlin/com/mapbox/maps/mapbox_maps/pigeons/CircleAnnotationMessenger.kt
@@ -415,6 +415,7 @@ interface _CircleAnnotationMessenger {
   fun update(managerId: String, annotation: CircleAnnotation, callback: (Result<Unit>) -> Unit)
   fun delete(managerId: String, annotation: CircleAnnotation, callback: (Result<Unit>) -> Unit)
   fun deleteAll(managerId: String, callback: (Result<Unit>) -> Unit)
+  fun deleteMulti(managerId: String, annotations: List<CircleAnnotation>, callback: (Result<Unit>) -> Unit)
   fun setCircleElevationReference(managerId: String, circleElevationReference: CircleElevationReference, callback: (Result<Unit>) -> Unit)
   fun getCircleElevationReference(managerId: String, callback: (Result<CircleElevationReference?>) -> Unit)
   fun setCircleSortKey(managerId: String, circleSortKey: Double, callback: (Result<Unit>) -> Unit)
@@ -562,6 +563,26 @@ interface _CircleAnnotationMessenger {
             val args = message as List<Any?>
             val managerIdArg = args[0] as String
             api.deleteAll(managerIdArg) { result: Result<Unit> ->
+              val error = result.exceptionOrNull()
+              if (error != null) {
+                reply.reply(wrapError(error))
+              } else {
+                reply.reply(wrapResult(null))
+              }
+            }
+          }
+        } else {
+          channel.setMessageHandler(null)
+        }
+      }
+      run {
+        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.mapbox_maps_flutter._CircleAnnotationMessenger.deleteMulti$separatedMessageChannelSuffix", codec)
+        if (api != null) {
+          channel.setMessageHandler { message, reply ->
+            val args = message as List<Any?>
+            val managerIdArg = args[0] as String
+            val annotationsArg = args[1] as List<CircleAnnotation>
+            api.deleteMulti(managerIdArg, annotationsArg) { result: Result<Unit> ->
               val error = result.exceptionOrNull()
               if (error != null) {
                 reply.reply(wrapError(error))