+++ example/lib/polyline_annotations_example.dart
@@ -20,6 +20,7 @@ class PolylineAnnotationExampleState extends State<PolylineAnnotationExample> {
   MapboxMap? mapboxMap;
   PolylineAnnotation? polylineAnnotation;
   PolylineAnnotationManager? polylineAnnotationManager;
+  List<PolylineAnnotation> annotations = [];
   int styleIndex = 1;
 
   _onMapCreated(MapboxMap mapboxMap) {
@@ -34,11 +35,16 @@ class PolylineAnnotationExampleState extends State<PolylineAnnotationExample> {
         positions.add(createRandomPositionList());
       }
 
-      polylineAnnotationManager?.createMulti(positions
-          .map((e) => PolylineAnnotationOptions(
-              geometry: LineString(coordinates: e),
-              lineColor: createRandomColor()))
-          .toList());
+      polylineAnnotationManager
+          ?.createMulti(positions
+              .map((e) => PolylineAnnotationOptions(
+                  geometry: LineString(coordinates: e),
+                  lineColor: createRandomColor()))
+              .toList())
+          .then((createdAnnotations) {
+        annotations =
+            createdAnnotations.whereType<PolylineAnnotation>().toList();
+      });
       polylineAnnotationManager?.tapEvents(onTap: (annotation) {
         // ignore: avoid_print
         print("onAnnotationClick, id: ${annotation.id}");
@@ -115,11 +121,25 @@ class PolylineAnnotationExampleState extends State<PolylineAnnotationExample> {
     );
   }
 
+  Widget _deleteMulti() {
+    return TextButton(
+        child: Text('delete 50 polyline annotations'),
+        onPressed: () async {
+          if (annotations.isNotEmpty) {
+            final toDelete = annotations.take(50).toList();
+            await polylineAnnotationManager?.deleteMulti(toDelete);
+            annotations.removeRange(
+                0, toDelete.length.clamp(0, annotations.length));
+          }
+        });
+  }
+
   Widget _deleteAll() {
     return TextButton(
       child: Text('delete all polyline annotations'),
       onPressed: () {
         polylineAnnotationManager?.deleteAll();
+        annotations.clear();
       },
     );
   }
@@ -132,7 +152,7 @@ class PolylineAnnotationExampleState extends State<PolylineAnnotationExample> {
     final List<Widget> listViewChildren = <Widget>[];
 
     listViewChildren.addAll(
-      <Widget>[_create(), _update(), _delete(), _deleteAll()],
+      <Widget>[_create(), _update(), _delete(), _deleteMulti(), _deleteAll()],
     );
 
     final colmn = Column(